* master:
  Return etherbi data divided by the token decimals
  Fixed the etherbi GQL tests
  Change the exchange fund flow to use float
  Use floats for the etherbi data

+++ lib/sanbase_web/graphql/schema/transaction_types.ex
@@ -17,7 +17,7 @@ defmodule SanbaseWeb.Graphql.TransactionTypes do
 
   object :exchange_transaction do
     field(:datetime, non_null(:datetime))
-    field(:transaction_volume, :decimal)
+    field(:transaction_volume, :float)
     field(:address, :string)
   end
 end
